requirements.txt
streamlit-chat==0.0.2.1
langchain==0.0.59
gpt-index==0.2.13
wikipedia==1.4.0
snowflake-sqlalchemy==1.4.4
